Split MarkV into multiple files
This CL breaks the monolithic markv_codec file into files for the base
class, encoder, decoder and logger.
